+++ prisma/migrations/20260115113635_add_created_at_to_categories_on_products/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "CategoriesOnProducts" ADD COLUMN     "createdAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP;